feat: fixture 분리
