+++ todos/tests/test_subtodo.py
@@ -1,52 +1,12 @@
 import pytest
+from django.contrib.auth import get_user_model
 from django.urls import reverse
 from rest_framework.test import APIClient
 
-from todos.models import Category, SubTodo, Todo, User
+from todos.models import SubTodo
 
+User = get_user_model()
 client = APIClient()
-
-
-@pytest.fixture
-def create_user(db):
-    user = User.objects.create_user(
-        username="testuser",
-        email="testuser@example.com",
-        password="testpassword",
-    )
-    return user
-
-
-@pytest.fixture
-def authenticated_client(create_user):
-    client.force_authenticate(user=create_user)
-    yield client
-    client.force_authenticate(user=None)  # logout
-
-
-@pytest.fixture
-def create_category(db, create_user):
-    category = Category.objects.create(
-        user_id=create_user,
-        title="Test Category",
-        color="#FFFFFF",
-        order="0|hzzzzz:",
-    )
-    return category
-
-
-@pytest.fixture
-def create_todo(db, create_user, create_category):
-    todo = Todo.objects.create(
-        user_id=create_user,
-        start_date="2024-08-01",
-        end_date="2024-08-30",
-        category_id=create_category,
-        content="Test Todo",
-        order="0|hzzzzz:",
-        is_completed=False,
-    )
-    return todo
 
 
 """
